+++ components/MenuList/MenuHeader.js
@@ -2,10 +2,9 @@
 import React from 'react';
 import styles from './Menu.module.scss';
 
-const MenuHeader = (props: { title: string, subtitle: string }) => (
+const MenuHeader = (props: { title: string }) => (
   <div className={styles.header}>
     <span className={styles.header__title}>{props.title}</span>
-    <span className={styles.header__subtitle}>{props.subtitle}</span>
   </div>
 );